% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kml_to_tiff.R
\name{kml_to_tiff}
\alias{kml_to_tiff}
\title{Convert a kml to a tif file}
\usage{
kml_to_tiff(folder_path, boundary, export_name)
}
\arguments{
\item{folder_path}{Path to the input folder with all of the .pngs of your .kml. Delete the "logo" .png ahead of time.}

\item{boundary}{A SpatVect around the actual data in your .kml. We recommend drawing this using ArcGIS around where you have data. If you're not interested in cropping your .kml, insert this as a blank shapefile the same size as your .kml.}

\item{export_name}{Path, and name, of the .tif you want to export from this function.}
}
\description{
This function turns a folder of .kml .pngs, specifically from SonarTRX, into a tif file.
}
